✨친구 검색 시, 로딩 스켈레톤 적용
