+++ src/app/(with-main-header)/friends/FriendsItem.tsx
@@ -1,21 +1,9 @@
-import { friendApi } from "@/api/friend/friend";
-import { revalidateTagAction } from "@/app/actions";
-import Button from "@/components/common/Button";
-import Modal from "@/components/common/Modal";
 import ProfileSummary from "@/components/common/ProfileSummary";
-import { useModalStore } from "@/store/modalStore";
+import { delay } from "@/utils/delay";
+import FriendDeleteButton from "./FriendDeleteButton";
 
-export default function FriendsItem({ friend }: { friend: friendType }) {
-  const { open, close } = useModalStore();
-
-  const deleteFriend = async () => {
-    const res = await friendApi.deleteFriend(friend.friendId);
-    if (res?.status === "성공") {
-      revalidateTagAction("friends");
-      close();
-    }
-  };
-
+export default async function FriendsItem({ friend }: { friend: friendType }) {
+  await delay(5000);
   return (
     <article className="h-19 lg:h-25 flex justify-between items-center">
       <ProfileSummary
@@ -24,17 +12,7 @@ export default function FriendsItem({ friend }: { friend: friendType }) {
         nickname={friend.nickname}
         description={friend.context}
       />
-      <Button onClick={() => open(`friend-delete-${friend.friendId}`)}>
-        삭제
-      </Button>
-      <Modal modalId={`friend-delete-${friend.friendId}`}>
-        <div className="flex flex-col items-center">
-          <p className="text-xl">친구를 삭제하시겠습니까?</p>
-          <div className="flex gap-5 mt-5">
-            <Button onClick={deleteFriend}>삭제하기</Button>
-          </div>
-        </div>
-      </Modal>
+      <FriendDeleteButton friend={friend} />
     </article>
   );
 }
